+++ src/extensions/mulle-sde/cmake/project/executable/cmake/CMakeDependenciesAndLibraries.cmake
@@ -2,7 +2,7 @@
 # Put your find_library() statements here to import other libraries
 #
 # Add OS specific dependencies to OS_SPECIFIC_LIBRARIES
-# Add all other dependencies (rest) to C_DEPENDENCIES_LIBRARIES
+# Add all other dependencies (rest) to DEPENDENCIES_LIBRARIES
 
 
 #
@@ -12,11 +12,10 @@
 #
 # === MULLE-SDE START ===
 
-include( .mulle-sde/etc/cmake/_CMakeDependencies.cmake)
-include( .mulle-sde/etc/cmake/_CMakeLibraries.cmake)
+include( _CMakeDependencies)
+include( _CMakeLibraries)
 
 # === MULLE-SDE END ===
 #
-
 
 # You can put more dependency definitions here.